Organises dirs and functions
