--- src/lib/functions.ts
@@ -1,7 +1,0 @@
-export type StylesCreator<Styles> = (theme: any) => Styles;
-
-export function createThemedStyles<Styles>(
-  stylesCreator: StylesCreator<Styles>
-) {
-  return (theme: any) => stylesCreator(theme);
-}
