+++ src/lib/functions/styles.ts
@@ -1,0 +1,9 @@
+import { Theme } from "@mui/material";
+
+export type StylesCreator<Styles> = (theme: Theme) => Styles;
+
+export function createThemedStyles<Styles>(
+  stylesCreator: StylesCreator<Styles>
+) {
+  return (theme: Theme) => stylesCreator(theme);
+}
